+++ packages/rettangoli-ui/src/components/tooltip/tooltip.store.js
@@ -6,7 +6,7 @@ export const selectViewData = ({ props }) => {
     open: !!props.open,
     x: props.x || 0,
     y: props.y || 0,
-    placement: props.placement || 'top',
+    place: props.place || 't',
     content: props.content || ''
   };
 }